% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{compute_exp_cost}
\alias{compute_exp_cost}
\title{Computes the expected posterior value of the objective function.
This function is private to this file.}
\usage{
compute_exp_cost(alphas, n, obj_fn, obj_fn_args)
}
\arguments{
\item{alphas}{A row vector from the matrix alphas given in the text.}

\item{n}{The number of new tracers to release.}

\item{obj_fn}{The objective function.}

\item{obj_fn_args}{Arguments to pass to the objective function.}
}
\value{
The expected posterior value of the objective function.
}
\description{
Computes the expected posterior value of the objective function.
This function is private to this file.
}

