% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_functions.R
\name{compute_prob}
\alias{compute_prob}
\title{Computes 1 - f(delta, alpha, beta), where f(delta, alpha, beta)
is the CDF of the beta distribution with shape parameters alpha and beta
evaluated at delta. This function is private to this file.}
\usage{
compute_prob(delta, alpha, beta)
}
\description{
Computes 1 - f(delta, alpha, beta), where f(delta, alpha, beta)
is the CDF of the beta distribution with shape parameters alpha and beta
evaluated at delta. This function is private to this file.
}

