% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_functions.R
\name{cv.beta}
\alias{cv.beta}
\title{Computes the coefficient of variance for a beta distribution with
parameters alpha and beta. This function is private to this file.}
\usage{
cv.beta(alpha, beta)
}
\arguments{
\item{alpha}{The first shape parameter of the beta distribution.}

\item{beta}{The second shape parameter of the beta distribution.}
}
\value{
The standard deviation for the beta distribution.
}
\description{
Computes the coefficient of variance for a beta distribution with
parameters alpha and beta. This function is private to this file.
}

