% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity_matrices.R
\name{generate_adj_mat_small_world}
\alias{generate_adj_mat_small_world}
\title{Generates a small world topology. This function is private to this file.}
\usage{
generate_adj_mat_small_world(n_sites, params)
}
\description{
Generates a small world topology. This function is private to this file.
}

