% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity_matrices.R
\name{generate_conn_mat}
\alias{generate_conn_mat}
\title{Generates a connectivity matrix as described in the manuscript text.}
\usage{
generate_conn_mat(n_sites, topology_params = NULL)
}
\arguments{
\item{n_sites}{The number of sites. This should be a perfect square
(4, 9, ...).}

\item{topology_params}{A list with elements p and nei, which are passed
through to igraph::watts.strogatz.game().}
}
\description{
Generates a connectivity matrix as described in the manuscript text.
}

