% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity_matrices.R
\name{generate_conn_mat_unif}
\alias{generate_conn_mat_unif}
\title{Divides the connection strengths among the edges as described in the text.
This function is private to this file.}
\usage{
generate_conn_mat_unif(adj_mat)
}
\description{
Divides the connection strengths among the edges as described in the text.
This function is private to this file.
}

