% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_functions.R
\name{mu.beta}
\alias{mu.beta}
\title{Compute the mean of a beta distribution with parameters alpha and beta. This
function is private to this file.}
\usage{
mu.beta(alpha, beta)
}
\arguments{
\item{alpha}{The first shape parameter of the beta distribution.}

\item{beta}{The second shape parameter of the beta distribution.}
}
\value{
The expected value for the beta distribution.
}
\description{
Compute the mean of a beta distribution with parameters alpha and beta. This
function is private to this file.
}

