% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_functions.R
\name{obj_fn_probabilities}
\alias{obj_fn_probabilities}
\title{Compute the objective function that is described in the text.}
\usage{
obj_fn_probabilities(alphas, obj_fn_args, by_site = FALSE)
}
\arguments{
\item{alphas}{An m-by-k matrix for a system with m origins and k
destinations. alphas[i,j] is x[i,j] as described in the manuscript text.

The following arguments should be in obj_fn_args.}

\item{delta}{The threshold for "irrelevant" connections.}

\item{epsilon}{The threshold for stopping.}

\item{pi}{The probability that a connection must be "irrelevant" for it to
be ignored.}
}
\value{
The value of the objective function.
}
\description{
Compute the objective function that is described in the text.
}

