% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity_matrices.R
\name{sample_conn_mat}
\alias{sample_conn_mat}
\title{\todo Add test that this is accurately simulating the matrix.}
\usage{
sample_conn_mat(conn_mat, origin, n)
}
\description{
\todo Add test that this is accurately simulating the matrix.
}

