% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_functions.R
\name{sd.beta}
\alias{sd.beta}
\title{Compute the standard deviation of a beta distribution with parameters alpha
and beta. This function is private to this file.}
\usage{
sd.beta(alpha, beta)
}
\arguments{
\item{alpha}{The first shape parameter of the beta distribution.}

\item{beta}{The second shape parameter of the beta distribution.}
}
\value{
The standard deviation for the beta distribution.
}
\description{
Compute the standard deviation of a beta distribution with parameters alpha
and beta. This function is private to this file.
}

